feat: refactor StampSVG options into separate file
- Moved STAMP_ASPECT_RATIO_OPTIONS and STAMP_FONT_OPTIONS to a new file, StampSVG.options.ts, for better organization and maintainability.
- Updated imports in StampSVG and related pages to reference the new options file.
